+++ contracts/PredictionMarket.sol
@@ -12,15 +12,18 @@ contract PredictionMarket {
     constructor(OutcomeOracle oracle_) {
         oracle = oracle_;
         commodities = new Commodity[](oracle.numOutcomes());
-        for (uint i = 0; i < oracle.numOutcomes(); i++) {
+        for (uint256 i = 0; i < oracle.numOutcomes(); i++) {
             commodities[i] = new Commodity(oracle.outcomes(i));
         }
     }
 
     function bet(address[] memory recipients) public payable {
         require(!oracle.decided(), "No betting after decision.");
-        require(recipients.length == oracle.numOutcomes(), "One recipient per outcome.");
-        for (uint i = 0; i < oracle.numOutcomes(); i++) {
+        require(
+            recipients.length == oracle.numOutcomes(),
+            "One recipient per outcome."
+        );
+        for (uint256 i = 0; i < oracle.numOutcomes(); i++) {
             commodities[i].mint(recipients[i], msg.value);
         }
     }
@@ -28,14 +31,14 @@ contract PredictionMarket {
     function payout(address payable recipient) public {
         require(oracle.decided(), "No payouts before decision.");
 
-        uint winnerValue = commodities[oracle.decision()].totalSupply();
-        uint totalValue = 0;
-        for (uint i = 0; i < oracle.numOutcomes(); i++) {
+        uint256 winnerValue = commodities[oracle.decision()].totalSupply();
+        uint256 totalValue = 0;
+        for (uint256 i = 0; i < oracle.numOutcomes(); i++) {
             totalValue += commodities[i].totalSupply();
         }
 
-        uint balance = commodities[oracle.decision()].balanceOf(msg.sender);
-        uint scaledBalance = (balance * totalValue) / winnerValue;
+        uint256 balance = commodities[oracle.decision()].balanceOf(msg.sender);
+        uint256 scaledBalance = (balance * totalValue) / winnerValue;
 
         commodities[oracle.decision()].burn(msg.sender, balance);
         recipient.transfer(scaledBalance);
